add routing for saved and rec pages
