+++ frontend/src/app/(dashboard)/recommendations/Recommendations.tsx
@@ -1,0 +1,13 @@
+'use client';
+
+const Recommendations = () => {
+  // placeholder
+  return (
+    <div>
+      <h1 className="text-3xl font-bold text-green-400 mb-2">Maya's Spotify Dashboard</h1>
+      <p className="text-md text-white/75 mt-1 mb-3">Discover your top tracks, playlists, genres and artists over time.</p>
+    </div>
+  );
+};
+
+export default Recommendations;
