+++ frontend/src/app/(dashboard)/saved/page.tsx
@@ -1,0 +1,14 @@
+'use client';
+
+import { Suspense } from 'react';
+import Saved from './Saved';
+
+const Page = () => {
+  return (
+    <Suspense fallback={<p>Loading Saved...</p>}>
+      <Saved />
+    </Suspense>
+  );
+};
+
+export default Page;
